// case-study-module5/src/app/model/facility.ts
export interface Facility {
  id?: number;
  facilityType?: string;
  name?: string;
  area?: number;
  cost?: number;
  maxPeople?: number;
  rentType?: string;
  standardRoom?: string;
  descriptionOtherConvenience?: string;
  poolArea?: number;
  numberFloor?: number;
  facilityFree?: string;
  image?: string;
}